test(cardano-services): consolidate entrypoint tests and rename to CLI
